Add main to crawler script
